+++ src/Pim/Bundle/CatalogBundle/Manager/MediaManager.php
@@ -10,6 +10,7 @@ use Pim\Bundle\CatalogBundle\Model\ProductMediaInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductValueInterface;
 use Pim\Bundle\CatalogBundle\Factory\MediaFactory;
+use Symfony\Component\HttpFoundation\File\Exception\FileNotFoundException;
 use Symfony\Component\HttpFoundation\File\File;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 
@@ -247,14 +248,17 @@ class MediaManager
 
     /**
      * Get the media, base64 encoded
-     *
      * @param ProductMediaInterface $media
      *
-     * @return string
+     * @return string|null the base 64 representation of the file media or null if the media has no file attached
+     *
+     * @throws FileNotFoundException in case the file of the media does not exist or is not readable
      */
     public function getBase64(ProductMediaInterface $media)
     {
-        return base64_encode(file_get_contents($this->getFilePath($media)));
+        $path = $this->getFilePath($media);
+
+        return $path !== null ? base64_encode(file_get_contents($this->getFilePath($media))) : null;
     }
 
     /**
@@ -308,16 +312,26 @@ class MediaManager
     }
 
     /**
-     * Read a file
+     * Get the file path of a media
+     *
      * @param ProductMediaInterface $media
      *
-     * @return string|null
+     * @return string|null the path of the media or null if the media has no file attached
+     *
+     * @throws FileNotFoundException in case the file of the media does not exist or is not readable
      */
     protected function getFilePath(ProductMediaInterface $media)
     {
         if ($this->fileExists($media)) {
-            return $this->uploadDirectory . DIRECTORY_SEPARATOR . $media->getFilename();
+            $path = $this->uploadDirectory . DIRECTORY_SEPARATOR . $media->getFilename();
+            if (!is_readable($path)) {
+                throw new FileNotFoundException($path);
+            }
+
+            return $path;
         }
+
+        return null;
     }
 
     /**