+++ CHANGELOG.md
@@ -110,6 +110,8 @@
 - PIM-3430: Fix doctrine issue on prices when skip an item during the product import
 - PIM-3358: Fix sprintf issue in an exception which prevents doctrine writer to deal with anything else than an object
 - PIM-3326: Fix mongo filters with multiples values and empty on MongoDB
+- PIM-3426: Fix common attributes edition on multi selects
+- PIM-3434: Fix bug in product media manager when file does not exist on the filesystem
 
 # 1.2.12 (2014-11-13)